Image RESTful controller + tests, also fixes

+++ services/ImageProcessingService.php
@@ -3,21 +3,24 @@
 namespace app\services;
 
 use \app\models\Image;
+use yii\web\UploadedFile;
 
 class ImageProcessingService
 {
 	private $watermarkFilepath = __DIR__ . '/../web/img/watermark.png';
 	private $uploadDir = __DIR__ . '/../web/img/upload/';
 
-	public function processImage($file) : Image
+	public function processImage(UploadedFile $file) : Image
 	{
-		if (!file_exists($file))
+	    $filename = $file->tempName;
+
+		if (!file_exists($filename))
 		{
 			throw new \Exception("File does not exists");
 		}
 
         $image = new \Imagick();
-		$image->readImage($file);
+		$image->readImage($filename);
 
 		// Create instance of the Watermark image
 		$watermark = new \Imagick();
@@ -33,38 +36,37 @@ class ImageProcessingService
 		// $image->compositeImage($watermark, \Imagick::COMPOSITE_OVER, $x, $y);
 
 		// Generate new name by file hash
-		$newname  = md5_file($file);
+		$newName  = md5_file($filename);
 
 		// Save original image
-		$origName = $newname . "." . $image->getImageFormat();
+		$origName = $newName . "." . $image->getImageFormat();
 		$previewName = 'pr_' . $origName;
 		$thumbnailName = 'tn_' . $origName;
 
 		$image->writeImage($this->uploadDir . $origName);
 
+        /**
+         * @var Image $imageAr
+         */
+        $imageAr = new Image();
+        $imageAr->scenario = 'create';
+
 		// Save preview image
-		$preview = $image->clone();
-		$preview->scaleImage(300, 300, true);
-		$preview->writeImage($this->uploadDir . $previewName);
+        $image->scaleImage(300, 300, true);
+        $image->writeImage($this->uploadDir . $previewName);
+        $imageAr->preview_width = $image->getImageWidth();
+        $imageAr->preview_height = $image->getImageHeight();
 
 		// Save thumb image
-		$thumbnail = $image->clone();
-		$thumbnail->scaleImage(150, 150, true);
-		$thumbnail->writeImage($this->uploadDir . $thumbnailName);
-
-		/**
-		 * @var \app\models\Image $imageAr
-		 */
-		$imageAr = new Image();
+        $image->scaleImage(150, 150, true);
+        $image->writeImage($this->uploadDir . $thumbnailName);
+        $imageAr->thumb_width = $image->getImageWidth();
+        $imageAr->thumb_height = $image->getImageHeight();
 		$imageAr->orig = $origName;
 		$imageAr->orig_width = $image->getImageWidth();
 		$imageAr->orig_height = $image->getImageHeight();
 		$imageAr->preview = $previewName;
-		$imageAr->preview_width = $preview->getImageWidth();
-		$imageAr->preview_height = $preview->getImageHeight();
 		$imageAr->thumb = $thumbnailName;
-		$imageAr->thumb_width = $thumbnail->getImageWidth();
-		$imageAr->thumb_height = $thumbnail->getImageHeight();
 
 		if (!$imageAr->save())
 		{
@@ -74,10 +76,15 @@ class ImageProcessingService
 		return $imageAr;
 	}
 
-	public function deleteImageFiles(\app\models\Image $image) : void
+	public function deleteImageFiles(Image $image) : void
 	{
-		unlink(\Yii::$app->params['uploadPath'] . $image->orig);
-		unlink(\Yii::$app->params['uploadPath'] . $image->preview);
-		unlink(\Yii::$app->params['uploadPath'] . $image->thumb);
+		if (file_exists(\Yii::$app->params['uploadPath'] . $image->orig)) { unlink(\Yii::$app->params['uploadPath'] . $image->orig); }
+		else { \Yii::warning("Missing original file for image $image->name ($image->id)"); }
+
+		if (file_exists(\Yii::$app->params['uploadPath'] . $image->preview)) { unlink(\Yii::$app->params['uploadPath'] . $image->preview); }
+		else { \Yii::warning("Missing preview file for image $image->name ($image->id)"); }
+
+		if (file_exists(\Yii::$app->params['uploadPath'] . $image->thumb)) { unlink(\Yii::$app->params['uploadPath'] . $image->thumb); }
+		else { \Yii::warning("Missing thumbnail file for image $image->name ($image->id)"); }
 	}
 }
